+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.1.0a12] - 2018-10-20
 ### Changed
  - `AirmoreSession::is_server_running` now uses socket instead of `requests`
  for faster, accurate and consistent check of Airmore server